requirements.txt
gtsam
loguru
seaborn
tqdm
pyyaml
evo
tabulate
plotly